Don't penalize new rubies
These `platforms` don't prevent the version restrictions because these
requirements are merged with the gemspec requirements later which
doesn't have platform restrictions. So bundling on ruby 2.4.2 still
restricts to faraday '0.9.2' which isn't very useful for testing new
faraday compatibility.

Instead just filter the lines completely based on ruby version.

Sorry this is a bit gnarly, but this is the most cross paltform and
backwards compatible way to get it working.